+++ R/funcs.R
@@ -278,3 +278,248 @@ replaceNULL <- function(x, replace = NA) {
     return(replace)
   x
 }
+
+
+chklevel <- function(x, length = 0, ci = TRUE, name = NULL, single = FALSE) {
+  if (!missing(single) && single)
+    length <- 1
+  ##
+  ## Check for levels of confidence interval / contour level
+  ##
+  if (is.null(name))
+    name <- deparse(substitute(x))
+  if (ci)
+    "level for confidence interval (range: 0-1)"
+  else
+    "contour levels (range: 0-1)"
+  ##
+  if (!is.numeric(x))
+    if (length && length(x) != length)
+    stop("Argument '", name, "' must be a numeric of length ", length, ".",
+         call. = FALSE)
+    else
+      stop("Argument '", name, "' must be numeric.",
+           call. = FALSE)
+  ##
+  if (length && length(x) != length)
+    stop("Argument '", name, "' must be a numeric of length ", length, ".",
+         call. = FALSE)
+  ##
+  if (any(x <= 0, na.rm = TRUE) | any(x >= 1, na.rm = TRUE))
+    stop("Argument '", name, "' must be a numeric between 0 and 1.",
+         call. = FALSE)
+  ##
+  invisible(NULL)
+}
+
+
+deprecated2 <- function(newvar, newmiss, oldvar, oldmiss, warn = FALSE) {
+  ##
+  new <- deparse(substitute(newvar))
+  old <- deparse(substitute(oldvar))
+  ##
+  if (newmiss & oldmiss)
+    return(newvar)
+  else if (!newmiss & oldmiss)
+    return(newvar)
+  else if (!newmiss & !oldmiss) {
+    if (warn)
+      warning("Deprecated argument '", old, "' ignored as ",
+              "'", new, "' is also provided.",
+              call. = FALSE)
+    return(newvar)
+  }
+  else if (newmiss & !oldmiss) {
+    if (warn)
+      warning("Use argument '", new, "' instead of '",
+              old, "' (deprecated).",
+              call. = FALSE)
+    return(oldvar)
+  }
+}
+
+
+formatN <- function(x, digits = 2, text.NA = "--", big.mark = "",
+                    format.whole.numbers = TRUE) {
+  
+  outdec <- options()$OutDec
+  
+  
+  if (format.whole.numbers) {
+    res <- format(ifelse(is.na(x),
+                         text.NA,
+                         formatC(x, decimal.mark = outdec,
+                                 format = "f", digits = digits,
+                                 big.mark = big.mark)
+                         )
+                  )
+  }
+  else {
+    res <- format(ifelse(is.na(x),
+                         text.NA,
+                  ifelse(is.wholenumber(x),
+                         x,
+                         formatC(x, decimal.mark = outdec,
+                                 format = "f", digits = digits,
+                                 big.mark = big.mark)
+                         )
+                  )
+                  )
+  }
+  ##
+  res <- rmSpace(res, end = TRUE)
+  ##
+  res
+}
+
+
+rmSpace <- function(x, end = FALSE, pat = " ") {
+  
+  if (!end) {
+    while (any(substring(x, 1, 1) == pat, na.rm = TRUE)) {
+      sel <- substring(x, 1, 1) == pat
+      x[sel] <- substring(x[sel], 2)
+    }
+  }
+  else {
+    last <- nchar(x)
+    
+    while (any(substring(x, last, last) == pat, na.rm = TRUE)) {
+      sel <- substring(x, last, last) == pat
+      x[sel] <- substring(x[sel], 1, last[sel] - 1)
+      last <- nchar(x)
+    }
+  }
+  
+  x
+}
+
+
+is.wholenumber <- function(x, tol = .Machine$double.eps^0.5) {
+  if (is.numeric(x))
+    res <- abs(x - round(x)) < tol
+  else
+    res <- NA
+  ##
+  res
+}
+
+
+formatCI <- function(lower, upper,
+                     bracket.left = gs("CIbracket"),
+                     separator = gs("CIseparator"),
+                     bracket.right,
+                     justify.lower = "right",
+                     justify.upper = justify.lower,
+                     lower.blank = gs("CIlower.blank"),
+                     upper.blank = gs("CIupper.blank"),
+                     ...
+                     ) {
+  
+  ## Change layout of CIs
+  ##
+  chkchar(bracket.left, length = 1)
+  chkchar(separator, length = 1)
+  if (!missing(bracket.right))
+    chkchar(bracket.right, length = 1)
+  ##
+  if (missing(bracket.left)) {
+    bracktype <- setchar(bracket.left, c("[", "(", "{", ""))
+    ##
+    if (bracktype == "[") {
+      bracketLeft <- "["
+      bracketRight <- "]"
+    }
+    else if (bracktype == "(") {
+      bracketLeft <- "("
+      bracketRight <- ")"
+    }
+    else if (bracktype == "{") {
+      bracketLeft <- "{"
+      bracketRight <- "}"
+    }
+    else if (bracktype == "") {
+      bracketLeft <- ""
+      bracketRight <- ""
+    }
+    ##
+    bracket.left <- bracketLeft
+  }
+  ##
+  if (missing(bracket.right))
+    bracket.right <- bracketRight
+  
+  format.lower <- format(lower, justify = justify.lower)
+  format.upper <- format(upper, justify = justify.upper)
+  ##
+  if (!lower.blank)
+    format.lower <- rmSpace(format.lower)
+  if (!upper.blank)
+    format.upper <- rmSpace(format.upper)
+  ##
+  if (separator == "-")
+    format.upper <-
+      paste0(ifelse(substring(format.upper, 1, 1) == "-", " ", ""),
+             format.upper)
+  ##
+  res <- ifelse(lower != "NA" & upper != "NA",
+                paste0(bracket.left,
+                       format.lower,
+                       separator,
+                       format.upper,
+                       bracket.right),
+                "")
+  ##
+  res
+}
+
+
+chkchar <- function(x, length = 0, name = NULL, nchar = NULL, single = FALSE) {
+  if (!missing(single) && single)
+    length <- 1
+  if (is.null(name))
+    name <- deparse(substitute(x))
+  ##
+  if (length && length(x) != length) {
+    if (length == 1)
+      stop("Argument '", name, "' must be a character string.",
+           call. = FALSE)
+    else
+      stop("Argument '", name, "' must be a character vector of length ",
+           length, ".",
+           call. = FALSE)
+  }
+  ##
+  if (length == 1) {
+    if (!is.null(nchar) && !(nchar(x) %in% nchar))
+      if (length(nchar) == 1 && nchar == 1)
+        stop("Argument '", name, "' must be a single character.",
+             call. = FALSE)
+      else
+        stop("Argument '", name, "' must be a character string of length ",
+             if (length(nchar) == 2)
+               paste0(nchar, collapse = " or ")
+             else
+               paste0(nchar, collapse = ", "),
+             ".",
+             call. = FALSE)
+  }
+  ##
+  if (!is.character(x))
+    stop("Argument '", name, "' must be a character vector.")
+  else {
+    if (!is.null(nchar) & any(!(nchar(x) %in% nchar)))
+      if (length(nchar) == 1 && nchar == 1)
+        stop("Argument '", name, "' must be a vector of single characters.",
+             call. = FALSE)
+      else
+        stop("Argument '", name, "' must be a character vector where ",
+             "each element has ",
+             if (length(nchar) == 2)
+               paste0(nchar, collapse = " or ")
+             else
+               paste0(nchar, collapse = ", "),
+             " characters.",
+             call. = FALSE)
+  }
+}