+++ R/league.crossnma.R
@@ -8,29 +8,32 @@
 #' @param x An object created with \code{\link{crossnma}}.
 #' @param median A logical indicating whether to use the median
 #'   (default) or mean to measure relative treatment effects.
-#' @param exp If TRUE (default), odds ratios are displayed. If FALSE,
-#'   log odds ratios will be presented.
+#' @param backtransf A logical indicating whether results should be
+#'   back transformed. If \code{backtransf = TRUE}, results for
+#'   \code{sm = "OR"} are presented as odds ratios rather than log
+#'   odds ratios, for example.
 #' @param order A vector of treatment names (character) representing
 #'   the order in which to display these treatments.
-#' @param cov1.value The participant covariate value of
-#'   \code{cov1} for which to report the results. Must be specified
-#'   for network meta-regression and when individual participant
-#'   dataset is used in the analysis. For dichotomous covariates, a
-#'   character of the level (used in the data) should be indicated.
-#' @param cov2.value The participant covariate value of
-#'   \code{cov2} for which to report the results. Must be specified
-#'   for network meta-regression and when individual participant
-#'   dataset is used in the analysis. For dichotomous covariates, a
-#'   character of the level (used in the data) should be indicated.
-#' @param cov3.value The participant covariate value of
-#'   \code{cov3} for which to report the results. Must be specified
-#'   for network meta-regression and when individual participant
-#'   dataset is used in the analysis. For dichotomous covariates, a
-#'   character of the level (used in the data) should be indicated.
+#' @param cov1.value The participant covariate value of \code{cov1}
+#'   for which to report the results. Must be specified for network
+#'   meta-regression and when individual participant dataset is used
+#'   in the analysis. For dichotomous covariates, a character of the
+#'   level (used in the data) should be indicated.
+#' @param cov2.value The participant covariate value of \code{cov2}
+#'   for which to report the results. Must be specified for network
+#'   meta-regression and when individual participant dataset is used
+#'   in the analysis. For dichotomous covariates, a character of the
+#'   level (used in the data) should be indicated.
+#' @param cov3.value The participant covariate value of \code{cov3}
+#'   for which to report the results. Must be specified for network
+#'   meta-regression and when individual participant dataset is used
+#'   in the analysis. For dichotomous covariates, a character of the
+#'   level (used in the data) should be indicated.
 #' @param digits The number of digits to be used when displaying the
 #'   results.
 #' @param direction The format to display the league table. Two
 #'   options "wide" (default) and "long".
+#' @param exp Deprecated argument (replaced by \code{backtransf}).
 #' @param \dots Additional arguments (ignored at the moment).
 #'
 #' @return
@@ -47,20 +50,21 @@
 #' # The data comes from randomized-controlled trials and
 #' # non-randomized studies (combined naively)
 #' head(ipddata) # participant-level data
-#' head(stddata) # study-level data
+#' stddata # study-level data
 #'
 #' # Create a JAGS model
 #' mod <- crossnma.model(treat, id, relapse, n, design,
-#'   prt.data = ipddata, std.data = stddata,sm="OR",
+#'   prt.data = ipddata, std.data = stddata,
 #'   reference = "A", trt.effect = "random", method.bias = "naive")
 #'
 #' # Fit JAGS model
-#' fit <-crossnma(mod,
-#' n.burnin =10,n.iter = 50, n.thin = 1, n.chains = 3)
+#' set.seed(1909)
+#' fit <- crossnma(mod, n.burnin = 10, n.iter = 50,
+#'   n.thin = 1, n.chains = 3)
 #'
 #' # Create league tables
-#' league(fit, exp = TRUE)                     #  wide format
-#' league(fit, exp = TRUE, direction = "long") #  long format
+#' league(fit)                     #  wide format
+#' league(fit, direction = "long") #  long format
 #'
 #' @method league crossnma
 #' @export
@@ -68,13 +72,14 @@
 
 league.crossnma <- function(x,
                             median = TRUE,
-                            exp = FALSE,
+                            backtransf = x$model$backtransf,
                             order = NULL,
                             cov1.value=NULL,
                             cov2.value=NULL,
                             cov3.value=NULL,
-                            digits = 2,
+                            digits = gs("digits"),
                             direction = "wide",
+                            exp = backtransf,
                             ...) {
 
   chkclass(x, "crossnma")
@@ -85,16 +90,28 @@ league.crossnma <- function(x,
   else
     central.tdcy <- "mean"
   ##
-  chklogical(exp)
-  ##
   chknumeric(digits, min = 0, length = 1)
   direction <- setchar(direction, c("wide", "long"))
-
-
+  ##
+  missing.backtransf <- missing(backtransf)
+  backtransf <-
+    deprecated2(backtransf, missing.backtransf, exp, missing(exp))
+  chklogical(backtransf)
+  if (backtransf & x$model$sm %in% c("MD", "SMD")) {
+    if (!missing.backtransf)
+      warning("No back transformation of results for (standardised) ",
+              "mean differences (argument 'backtransf').")
+    backtransf <- FALSE
+  }
+  ##
+  exp <- backtransf
+  quant <- x$model$quantiles
+  
+  
   if (!is.null(x$model$covariate) & is.null(cov1.value))
     stop("cov1.value must be specified for network meta-regression")
-
-
+  
+  
   ## Bind variables to function
   trt <- Treatment <- Comparator <- cov.ref <- NULL
   #samples <- as.mcmc(x$jagsfit)
@@ -479,21 +496,31 @@ league.crossnma <- function(x,
       eval(call(fct,(call(trans, x)), arg))
   }
   ##
-  exp.mean <- function(x) calc.report(x, "mean", trans = "exp")
-  exp.median <- function(x) calc.report(x, "median", trans = "exp")
-  exp.sd <- function(x) calc.report(x, "sd", trans = "exp")
-  exp.lci <- function(x) calc.report(x, "quantile", arg = 0.025, trans = "exp")
-  exp.uci <- function(x) calc.report(x, "quantile", arg = 0.975, trans = "exp")
+  exp.mean <- function(x)
+    calc.report(x, "mean", trans = "exp")
+  exp.median <- function(x)
+    calc.report(x, "median", trans = "exp")
+  exp.sd <- function(x)
+    calc.report(x, "sd", trans = "exp")
+  exp.lcl <- function(x)
+    calc.report(x, "quantile", arg = quant[1], trans = "exp")
+  exp.ucl <- function(x)
+    calc.report(x, "quantile", arg = quant[3], trans = "exp")
   ##
-  id.mean <- function(x) calc.report(x, "mean", trans = "identity")
-  id.median <- function(x) calc.report(x, "median", trans = "identity")
-  id.sd <- function(x) calc.report(x, "sd", trans = "identity")
-  id.lci <- function(x) calc.report(x, "quantile", arg = 0.025, trans = "identity")
-  id.uci <- function(x) calc.report(x, "quantile", arg = 0.975, trans = "identity")
+  id.mean <- function(x)
+    calc.report(x, "mean", trans = "identity")
+  id.median <- function(x)
+    calc.report(x, "median", trans = "identity")
+  id.sd <- function(x)
+    calc.report(x, "sd", trans = "identity")
+  id.lcl <- function(x)
+    calc.report(x, "quantile", arg = quant[1], trans = "identity")
+  id.ucl <- function(x)
+    calc.report(x, "quantile", arg = quant[3], trans = "identity")
   ##
   colvals <- function(dmat, b.col=1, paste = TRUE) {
     ## Bind variables to function
-    key <- value <- trt <- estimate <- lci <- uci <- result <- NULL
+    key <- value <- trt <- estimate <- lcl <- ucl <- result <- NULL
 
     base <- colnames(dmat)[b.col]
 
@@ -532,60 +559,57 @@ league.crossnma <- function(x,
     }
     ##
     if (exp) {
-      tmp.lci <- dmat2 %>%
-        summarise_all(list(lci = exp.lci)) %>% gather() %>%
-        rename(trt = key, lci = value) %>%
-        mutate(trt = sub("_lci", "", trt))
+      tmp.lcl <- dmat2 %>%
+        summarise_all(list(lcl = exp.lcl)) %>% gather() %>%
+        rename(trt = key, lcl = value) %>%
+        mutate(trt = sub("_lcl", "", trt))
       ##
-      tmp.uci <- dmat2 %>%
-        summarise_all(list(uci = exp.uci)) %>% gather() %>%
-        rename(trt = key, uci = value) %>%
-        mutate(trt = sub("_uci", "", trt))
+      tmp.ucl <- dmat2 %>%
+        summarise_all(list(ucl = exp.ucl)) %>% gather() %>%
+        rename(trt = key, ucl = value) %>%
+        mutate(trt = sub("_ucl", "", trt))
       ##
       null.value <- 1
     }
     else {
-      tmp.lci <- dmat2 %>%
-        summarise_all(list(lci = id.lci)) %>% gather() %>%
-        rename(trt = key, lci = value) %>%
-        mutate(trt = sub("_lci", "", trt))
+      tmp.lcl <- dmat2 %>%
+        summarise_all(list(lcl = id.lcl)) %>% gather() %>%
+        rename(trt = key, lcl = value) %>%
+        mutate(trt = sub("_lcl", "", trt))
       ##
-      tmp.uci <- dmat2 %>%
-        summarise_all(list(uci = id.uci)) %>% gather() %>%
-        rename(trt = key, uci = value) %>%
-        mutate(trt = sub("_uci", "", trt))
+      tmp.ucl <- dmat2 %>%
+        summarise_all(list(ucl = id.ucl)) %>% gather() %>%
+        rename(trt = key, ucl = value) %>%
+        mutate(trt = sub("_ucl", "", trt))
       ##
       null.value <- 0
     }
 
 
-    ## create C-style formatting string from the digits parameter
+    ## Create C-style formatting string from the digits parameter
+    ##
     fmt <- paste0("%.", digits, "f")
 
     if (paste) {
-      tmp1 <- left_join(tmp.estimate, tmp.lci, by = "trt") %>%
-        left_join(tmp.uci, by = "trt") %>%
+      tmp1 <- left_join(tmp.estimate, tmp.lcl, by = "trt") %>%
+        left_join(tmp.ucl, by = "trt") %>%
         mutate(result = paste(sprintf(fmt, estimate),
-                              " (",
-                              sprintf(fmt, lci),
-                              " to ",
-                              sprintf(fmt, uci),
-                              ")", sep="")) %>%
+                              formatCI(round(lcl, digits),
+                                       round(ucl, digits)))) %>%
         select(trt, result)
       colnames(tmp1)[2] <-
         as.character(tmp.estimate %>% filter(estimate == null.value) %>%
                      select(trt))
     }
     else {
-      tmp1 <- left_join(tmp.estimate, tmp.lci, by = "trt") %>%
-        left_join(tmp.uci, by = "trt")
+      tmp1 <- left_join(tmp.estimate, tmp.lcl, by = "trt") %>%
+        left_join(tmp.ucl, by = "trt")
       colnames(tmp1)[2] <- central.tdcy
     }
-
     tmp1
   }
-
-
+  
+  
   ##
   ##
   ## Return league table
@@ -626,16 +650,42 @@ league.crossnma <- function(x,
     longtable <- tmp2.list %>%
       bind_rows() %>%
       mutate(Treatment = trt,
-             Comparator = rep(trt.names, each=length(trt.names))) %>%
+             Comparator =
+               rep(trt.names, each = length(trt.names))) %>%
       select(Treatment, Comparator, everything(), -trt)
     ##
-    class(longtable) <- c("league.crossnma", class(longtable))
+    fmt <- paste0("%.", digits, "f")
+    ##
+    if (median)
+      longtable$median <-
+        sprintf(fmt, round(longtable$median, digits))
+    else
+      longtable$mean <-
+        sprintf(fmt, round(longtable$mean, digits))
+    ##
+    longtable$lcl <-
+      sprintf(fmt, round(longtable$lcl, digits))
+    ##
+    longtable$ucl <-
+      sprintf(fmt, round(longtable$ucl, digits))
+    ##
+    longtable <-
+      longtable[longtable$Treatment != longtable$Comparator, ,
+                drop = FALSE]
+    ##
+    longtable <- as.matrix(longtable)
+    row.names(longtable) <- rep("", nrow(longtable))
+    cl <- colnames(longtable)
+    cl[cl %in% c("lcl", "ucl")] <- paste0(100 * quant[c(1, 3)], "%")
+    colnames(longtable) <- cl
+    ##
+    class(longtable) <- c("league.crossnma", "matrix")
     attr(longtable, "direction") <- direction
     ##
     return(longtable)
   }
-
-
+  
+  
   invisible(NULL)
 }
 
@@ -668,10 +718,8 @@ print.league.crossnma <- function(x, ...) {
     colnames(x) <- rep("", ncol(x))
     prmatrix(x, quote = FALSE, right = TRUE)
   }
-  else if (attr(x, "direction") == "long") {
-    class(x) <- "data.frame"
-    print(x)
-  }
-
+  else if (attr(x, "direction") == "long")
+    prmatrix(x, quote = FALSE, right = TRUE, ...)
+  
   invisible(NULL)
 }